requirements.txt
boto3==1.7.64
botocore==1.10.64
certifi==2018.4.16
chardet==3.0.4
docutils==0.14
idna==2.7
jmespath==0.9.3
neo4j-driver==1.6.1
neotime==1.0.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
requests-toolbelt==0.8.0
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
